chore(wip): consolidate in-flight master work onto a single commit
Bundles uncommitted user/agent WIP that had accumulated across multiple
in-flight efforts. Verified `cargo test -p nmp-core --lib` passes (396/396)
and the M2 subscription-compilation audit suite is green.

- kernel/outbox + kernel/requests/profile: split discovery vs content
  relays — profile-claim (kind:0) lookups now route through the indexer
  set ONLY (purplepag.es), never the content relay (damus.io). New
  helper Kernel::author_indexer_relays(); kernel/tests.rs tightened to
  assert single-REQ-to-indexer; kernel/profile_claim_tests.rs adds TDD
  coverage for the batched-by-pubkey claim path
- relay_worker: drop redundant `as u64` cast in jittered_backoff (clippy)
- nmp-testing/store_harness: replace UUID dep with pid+counter+nanos
  temp-dir naming for parallel-test isolation; matches the workspace
  decision to keep nmp-testing dep-light
- nmp-testing/m2_subscription_compilation_audit: rename
  unknown_author_falls_back_to_indexer →
  unknown_author_with_no_app_relays_becomes_unroutable to match T134's
  routing-rules clarification (indexer is discovery-only; never a
  content fallback)
- nmp-substrate-types/store/mod.rs: re-add file (it was deleted-by-us
  in a prior aborted merge; the content was already present on disk)
- planner/selection.rs: trivial std::collections::BTreeSet::new() →
  BTreeSet::new() (import already in scope; reduces noise)

+++ crates/nmp-core/src/relay_worker/mod.rs
@@ -110,7 +110,7 @@ pub(crate) fn jittered_backoff(base: Duration, url: &str) -> Duration {
     let hash = url
         .bytes()
         .fold(0u64, |acc, b| acc.wrapping_mul(31).wrapping_add(b as u64));
-    let jitter_ms = (hash % 5000) as u64; // 0–4999 ms spread
+    let jitter_ms = hash % 5000; // 0–4999 ms spread
     base + Duration::from_millis(jitter_ms)
 }
